Modularize files
- Moved Connection class to connection.js
- Moved Statement class to statement.js
- Moved DBPoolConnection class to dbPoolConnection.js
- Instead of hard coding constants iterate over idb provided constants
